+++ constants.h
@@ -1,7 +1,7 @@
 #ifndef _CONSTANTS_H_
 #define _CONSTANTS_H_
 
-const size_t	DefaultVelocity=100;
+const float		DefaultVelocity=100;
 const size_t	DefaultScores = 500;
 const float		PI=3.141592f;
 const size_t	DefaultNumPoints	= 50;
